refactor(collection): split skill assets out of io.ts
io.ts had become the glue between two unrelated concerns: record I/O (list /
read / write / delete / generateItemId) and skill-owned assets (custom-view
files plus the action seed-prompt assembly). Only the second half needs to know
where skills live on disk, and it was the only reason the storage layer imported
skillsStagingDir at all.

Moving it out drops io.ts from 515 to 239 lines and takes its host coupling from
four symbols to exactly three — getWorkspaceRoot, log, publishCollectionChange —
which is the set #2248 measured as DI-able. The storage layer no longer knows
the skill layout.

Callers are untouched: the barrel already did `export * from "./io"`, so adding
`export * from "./skillAssets"` next to it keeps every import path working.

Step 1 of #2248. It is deliberately independent of the packaging question still
open there (A/B/C) — this split is needed under any of them.
